refactor: extract TrainMovementManager from Train.java
Move moveLinkers() (193 lines), clearReservations(), and crash()
to a new TrainMovementManager class. Train.java delegates to it.

Train.java: 1318 → 1055 lines (-263)

Closes #87
