+++ src/main/java/letrain/vehicle/impl/rail/Train.java
@@ -39,7 +39,7 @@ import org.slf4j.LoggerFactory;
 @JsonIdentityInfo(generator = ObjectIdGenerators.IntSequenceGenerator.class, property = "@id")
 @com.fasterxml.jackson.annotation.JsonIgnoreProperties(ignoreUnknown = true)
 public class Train implements Trailer<RailTrack>, Renderable, Transportable {
-    private static final int CRASH_SPEED_THRESHOLD = 5;
+    static final int CRASH_SPEED_THRESHOLD = 5;
     private static final int MAX_LOADING_COUNT = 80; // 4.0 seconds at 20fps per wagon
     private static final Logger log = LoggerFactory.getLogger(Train.class);
     @com.fasterxml.jackson.annotation.JsonProperty("linkers")
@@ -67,6 +67,7 @@ public class Train implements Trailer<RailTrack>, Renderable, Transportable {
     @com.fasterxml.jackson.annotation.JsonUnwrapped
     private TrainLogisticsManager logisticsManager = new TrainLogisticsManager();
 
+    private final transient TrainMovementManager movementManager = new TrainMovementManager(this);
     private transient TrainSafetyManager safetyManager = new TrainSafetyManager(this);
     int railStationId = 0;
     private boolean stalled = false;
@@ -631,271 +632,7 @@ public class Train implements Trailer<RailTrack>, Renderable, Transportable {
     }
 
     public boolean moveLinkers(boolean isNormalSense) {
-        if (linkers.isEmpty()) {
-            return false;
-        }
-
-        // Pass 1: Verify all linkers can move to their next tracks
-        List<Track> targetTracks = new ArrayList<>();
-        Map<Linker, Track> currentTracks = new HashMap<>(); // next -> currentTrack
-        Map<Linker, Dir> entryDirsMap = new HashMap<>();
-        Map<Linker, Linker> occupyingLinkerMap = new HashMap<>();
-
-        List<Linker> movingOrder = new ArrayList<>();
-        if (isNormalSense) {
-            movingOrder.addAll(linkers);
-        } else {
-            Iterator<Linker> it = linkers.descendingIterator();
-            while (it.hasNext()) {
-                movingOrder.add(it.next());
-            }
-        }
-
-        Linker firstLinker = movingOrder.get(0);
-        Linker lastLinker = movingOrder.get(movingOrder.size() - 1);
-
-        for (Linker linkerToMove : movingOrder) {
-            Track currentTrack = linkerToMove.getTrack();
-            if (currentTrack == null) {
-                return false;
-            }
-
-            Dir exitDir = linkerToMove.getDir();
-            Track nextTrackOfLinker = currentTrack.getConnected(exitDir);
-
-            if (nextTrackOfLinker == null) {
-                log.debug("Pass 1: nextTrack is null for {}", linkerToMove);
-                clearReservations(targetTracks);
-                return false;
-            }
-
-            Linker occupyingL = nextTrackOfLinker.getLinker();
-            if (occupyingL != null) {
-
-                if (occupyingL.getTrain() != this) {
-                    int speed = getSpeed();
-
-                    if (Math.abs(speed) >= CRASH_SPEED_THRESHOLD) {
-                        crash(occupyingL, speed);
-                    } else {
-                        letrain.map.Point collisionPos = occupyingL.getPosition();
-                        notifyContact(collisionPos, speed);
-                        getTractors().forEach(t -> {
-                            t.setCurrentSpeed(0);
-                            t.setTargetSpeed(0);
-                        });
-                        this.setStalled(true);
-                        Train otherTrain = occupyingL.getTrain();
-                        if (otherTrain != null) {
-                            otherTrain.notifyContact(collisionPos, speed);
-                        }
-                    }
-                    clearReservations(targetTracks);
-                    return false;
-                }
-            }
-
-            Dir entryDirOfLinker = linkerToMove.getDir().inverse();
-            if (occupyingL == null || occupyingL.getTrain() != this) {
-                if (!nextTrackOfLinker.canEnter(entryDirOfLinker, linkerToMove)) {
-                    clearReservations(targetTracks);
-                    return false;
-                }
-            }
-
-            currentTracks.put(linkerToMove, currentTrack);
-            entryDirsMap.put(linkerToMove, entryDirOfLinker);
-            occupyingLinkerMap.put(linkerToMove, occupyingL);
-
-            nextTrackOfLinker.setReservation(linkerToMove);
-            targetTracks.add(nextTrackOfLinker);
-        }
-
-        // Pass 2: Actually move the linkers
-        for (int i = 0; i < movingOrder.size(); i++) {
-            Linker linkerToMove = movingOrder.get(i);
-            Track currentTrack = currentTracks.get(linkerToMove);
-            Track nextTrackOfLinker = targetTracks.get(i);
-            Dir entryDirOfLinker = entryDirsMap.get(linkerToMove);
-
-            Sensor sensorExit = currentTrack.getSensor();
-            if (sensorExit != null && linkerToMove == lastLinker) {
-                sensorExit.onExitTrain(this);
-            }
-            if (currentTrack.getSemaphore() != null && linkerToMove == lastLinker) {
-                currentTrack.getSemaphore().onExitTrain(this);
-            }
-            if (currentTrack instanceof ForkRailTrack && linkerToMove == lastLinker) {
-                ((ForkRailTrack) currentTrack).onExitTrain(this);
-            }
-
-            linkerToMove.setPreviousTrack(currentTrack);
-            linkerToMove.setPreviousDir(linkerToMove.getDir());
-            currentTrack.removeLinker();
-            if (!nextTrackOfLinker.enterLinkerFromDir(entryDirOfLinker, linkerToMove)) {
-                // Rollback: the linker was removed from currentTrack but could not enter
-                // nextTrack (e.g. another linker from the same train still occupies it).
-                // Restore the linker to its previous track to avoid linkers with null track.
-                linkerToMove.setTrack(currentTrack);
-                currentTrack.setLinker(linkerToMove);
-                linkerToMove.setPreviousTrack(null);
-                linkerToMove.setPreviousDir(null);
-                clearReservations(targetTracks);
-                return false;
-            }
-            linkerToMove.setRailsSinceStop(linkerToMove.getRailsSinceStop() + 1);
-
-            nextTrackOfLinker.setReservation(null);
-
-            Sensor sensorEnter = nextTrackOfLinker.getSensor();
-            if (sensorEnter != null && linkerToMove == firstLinker) {
-                sensorEnter.onEnterTrain(this);
-            }
-            if (nextTrackOfLinker.getSemaphore() != null && linkerToMove == firstLinker) {
-                nextTrackOfLinker.getSemaphore().onEnterTrain(this);
-            }
-            if (nextTrackOfLinker instanceof ForkRailTrack && linkerToMove == firstLinker) {
-                ((ForkRailTrack) nextTrackOfLinker).onEnterTrain(this);
-            }
-        }
-
-        // After a successful move, immediately check whether the next cell
-        // ahead is blocked by another train. If so, trigger collision NOW
-        // instead of waiting for the next advance() cycle (which would add
-        // up to 2.5s of delay before the crash sound plays).
-        Track currentFirstTrack = firstLinker.getTrack();
-        if (currentFirstTrack == null) {
-            log.warn("First linker has no track after move sequence — cannot check next cell for collisions");
-            clearReservations(targetTracks);
-            return false;
-        }
-        Track nextAfterMove = currentFirstTrack.getConnected(firstLinker.getDir());
-        if (nextAfterMove != null) {
-            Linker blockingLinker = nextAfterMove.getLinker();
-            if (blockingLinker != null && blockingLinker.getTrain() != this) {
-                int speed = getSpeed();
-                    if (Math.abs(speed) >= CRASH_SPEED_THRESHOLD) {
-                        crash(blockingLinker, speed);
-                        this.setStalled(true);
-                    } else {
-                        letrain.map.Point collisionPos = blockingLinker.getPosition();
-                    notifyContact(collisionPos, speed);
-                    getTractors().forEach(t -> {
-                        t.setCurrentSpeed(0);
-                        t.setTargetSpeed(0);
-                        if (t instanceof Locomotive) {
-                            ((Locomotive) t).setForceIdleSound(true);
-                        }
-                    });
-                    this.setStalled(true);
-                    Train otherTrain = blockingLinker.getTrain();
-                    if (otherTrain != null) {
-                        otherTrain.notifyContact(collisionPos, speed);
-                    }
-                }
-            }
-        } else {
-            // Train has reached a dead-end (no track connected ahead).
-            // Treat as collision: crash at high speed, contact at low speed.
-            int speed = getSpeed();
-            letrain.map.Point impactPos = firstLinker.getPosition();
-            if (Math.abs(speed) >= 5) {
-                // High-speed crash into dead-end -> destruction of THIS train
-                boolean alreadyDestroying = false;
-                for (Linker l : getLinkers()) {
-                    if (l instanceof Destructible && ((Destructible) l).isDestroying()) {
-                        alreadyDestroying = true;
-                        break;
-                    }
-                }
-                if (!alreadyDestroying) {
-                    notifyCrash(impactPos, speed);
-                    getLinkers().forEach(l -> {
-                        if (l instanceof Locomotive) {
-                            ((Locomotive) l).setCurrentSpeed(0);
-                            ((Locomotive) l).setTargetSpeed(0);
-                            ((Locomotive) l).setAcousticSpeedSignal(-1);
-                            ((Locomotive) l).setEngineTransitioning(false);
-                            ((Locomotive) l).setForceIdleSound(true);
-                        }
-                        l.destroy();
-                    });
-                    this.setStalled(true);
-                }
-            } else {
-                // Low-speed contact with dead-end -> stop without destruction
-                notifyContact(impactPos, speed);
-                getTractors().forEach(t -> {
-                    t.setCurrentSpeed(0);
-                    t.setTargetSpeed(0);
-                    if (t instanceof Locomotive) {
-                        ((Locomotive) t).setAcousticSpeedSignal(-1);
-                        ((Locomotive) t).setEngineTransitioning(false);
-                        ((Locomotive) t).setForceIdleSound(true);
-                    }
-                });
-                this.setStalled(true);
-            }
-        }
-
-        return true;
-    }
-
-    private void clearReservations(List<Track> reservedTracks) {
-        for (Track t : reservedTracks) {
-            t.setReservation(null);
-        }
-    }
-
-    private void crash(Linker linker, int speed) {
-        letrain.map.Point crashPos = linker.getPosition();
-        // Only trigger crash logic if this train isn't already destroying
-        boolean alreadyDestroying = false;
-        for (Linker l : getLinkers()) {
-            if (l instanceof Destructible && ((Destructible) l).isDestroying()) {
-                alreadyDestroying = true;
-                break;
-            }
-        }
-
-        if (!alreadyDestroying) {
-            notifyCrash(crashPos, speed);
-            getLinkers().forEach(l -> {
-                if (l instanceof Locomotive) {
-                    ((Locomotive) l).setCurrentSpeed(0);
-                    ((Locomotive) l).setTargetSpeed(0);
-                    ((Locomotive) l).setForceIdleSound(true);
-                }
-                l.destroy();
-            });
-            this.setStalled(true);
-        }
-
-        // Also handle the other linker/train
-        if (linker.getTrain() != null) {
-            boolean otherAlreadyDestroying = false;
-            for (Linker l : linker.getTrain().getLinkers()) {
-                if (l instanceof Destructible && ((Destructible) l).isDestroying()) {
-                    otherAlreadyDestroying = true;
-                    break;
-                }
-            }
-            if (!otherAlreadyDestroying) {
-                linker.getTrain().notifyCrash(crashPos, speed);
-                linker.getTrain().getLinkers().forEach(l -> {
-                    if (l instanceof Locomotive) {
-                        ((Locomotive) l).setCurrentSpeed(0);
-                        ((Locomotive) l).setTargetSpeed(0);
-                        ((Locomotive) l).setForceIdleSound(true);
-                    }
-                    l.destroy();
-                });
-                linker.getTrain().setStalled(true);
-            }
-        } else {
-            log.info("crash: Destroying loose linker {} at crash position {}", linker, crashPos);
-            linker.destroy();
-        }
+        return movementManager.moveLinkers(isNormalSense);
     }
 
     @JsonIgnore
